Fix youtube size on mobile and landing page icon tweak
This commit adds an include for youtube videos that wraps the iframes in
properly sized divs so that they are displayed with the appropriate
width for the users device - previously they'd been setting a hard 560px
width and were subsequently causing posts rendered on mobile to overflow
to the right.

This commit also tweaks the author buttons on the landing page to use a
non-breaking space between the fa icon and button name.

Signed-off-by: John McCabe <john@johnmccabe.net>