Render hero full-height pre mid-point
This commit alters the rendering behaviour of the hero so that it
renders full height pre- the 620px midpoint. This avoids the overflow
issue seen on mobile with post titles/descriptions.

NOTE - some common sense still needs to be taken with
titles/descriptions to ensure they are concise and not overly long.

Signed-off-by: John McCabe <john@johnmccabe.net>