Add homepage social media preview image
LinkedIn was finding the `og:image` tag but deciding not to use the image.  Further investigation and experimentation has found that the image originally used for the `og:image` tag was not of appropriate dimensions for LinkedIn.  The image requirements specific to the LinkedIn sharing module are stated to be:

Max file size: 5 MB
Minimum image dimensions: 1200 (w) x 627 (h) pixels
Recommended ratio: 1.91:1

https://www.linkedin.com/help/linkedin/answer/46687/making-your-website-shareable-on-linkedin?lang=en

This change adds an image that is 1435 x 750.

Signed-off-by: Richard Gee <richard@technologee.co.uk>